[REFACTOR] fix routing problem
